fix: rename apis index to axios
